Raised error for 53.5,138.9 after May 15
Not astronomical twilight time could be calculated at this time of
year, for Siberia. Changed code so that only a failure to find the
sunrise or sunset raises an error. The Twilight times can now be set to
nil, if they don’t exist on the same day as the sunrise or sunset.